refactor(ui): extract presentation helpers from lib/api.ts (#1965)
First incremental step toward the broader `ui/lib/api.ts` decomposition
tracked in #1965. The file is currently 2,134 lines — the largest TS
module in the dashboard — and it bundles three concerns into one
module:

1. Type / interface declarations (~110 of them).
2. The `api` object with its ~100 method shorthands.
3. Standalone presentation helpers (`severityColor`, `severityDot`,
   `formatDate`, `isConfigured`).

#3 is the smallest, most isolated cluster and the safest to move first.
This PR extracts those four helpers into a new `ui/lib/api-format.ts`
and re-exports them from `lib/api.ts` so every existing
`import { severityColor } from "@/lib/api"` keeps working unchanged.
50+ call sites stay untouched.

Public surface unchanged — `tsc --noEmit` clean, `eslint` clean,
`vitest` 144/144. Future PRs in the #1965 series will lift the type
declarations and per-domain method clusters out of `api.ts` the same
way.

Refs #1965 (one phase of that work).
